refactor: split story-orchestrator.ts into story-orchestrator/ directory
1473-line file (2.4x the 600-line hard limit) split into 8 focused modules
under src/execution/story-orchestrator/ with a barrel index.ts. No logic
changes — pure structural refactor. File-sizes baseline updated 17→16.

- types.ts        (204 lines) — interfaces, constants, CANONICAL_ORDER
- phase-state.ts  (42 lines)  — isSlot, setPhase, collectOrderedPhases
- phase-eval.ts   (160 lines) — phasePassed, extractPhaseFindings, gateFailureKeys
- rectification.ts (244 lines) — runRectification, gatherRectificationFindings
- review-decision.ts (142 lines) — toReviewDecisionPayload, refreshing review logs
- run-phase.ts    (235 lines) — _storyOrchestratorDeps, runPhase, refreshReviewInputForDispatch
- execution-plan.ts (336 lines) — ExecutionPlan class
- builder.ts      (134 lines) — StoryOrchestratorBuilder class
- index.ts        (29 lines)  — public API barrel
