+++ scripts/baselines/file-sizes-baseline.json
@@ -1,5 +1,5 @@
 {
-  "updatedAt": "2026-06-19T03:04:21.301Z",
+  "updatedAt": "2026-06-19T03:19:35.176Z",
   "byFile": {
     "src/agents/acp/adapter.ts": 629,
     "src/agents/acp/spawn-client.ts": 737,
@@ -7,7 +7,6 @@
     "src/config/runtime-types.ts": 626,
     "src/context/engine/providers/code-neighbor.ts": 613,
     "src/execution/post-run.ts": 607,
-    "src/execution/story-orchestrator.ts": 1473,
     "src/execution/unified-executor.ts": 689,
     "src/operations/call.ts": 631,
     "src/prompts/builders/rectifier-builder.ts": 902,
