+++ src/execution/story-orchestrator/index.ts
@@ -1,0 +1,29 @@
+// Public API barrel — re-exports every symbol that story-orchestrator.ts previously exported.
+// Internal files may import each other directly via relative leaf paths within this directory.
+
+export { StoryOrchestratorBuilder } from "./builder";
+export { ExecutionPlan } from "./execution-plan";
+export { extractPhaseFindings, gateFailureKeys, orderGateLast, phasesToRevalidate } from "./phase-eval";
+export { runRectification } from "./rectification";
+export { _storyOrchestratorDeps, refreshReviewInputForDispatch, withIncreasingFailuresBail } from "./run-phase";
+export {
+  EXHAUSTED_EXIT_REASONS,
+  CANONICAL_ORDER,
+  PHASE_KIND_TO_STATE_KEY,
+  STRATEGY_TO_REVALIDATION_PHASES,
+  TDD_OP_NAMES,
+  STRICT_VERDICT_PHASE_NAMES,
+} from "./types";
+export type {
+  AnySlot,
+  DroppedFindingSummary,
+  InternalBuildState,
+  InternalPhase,
+  OrchestratorSlot,
+  PhaseKind,
+  RectificationOverrides,
+  RectificationPhaseOptions,
+  RectificationResult,
+  ReviewDecisionPayload,
+  StoryOrchestratorResult,
+} from "./types";
